+++ database/migrations/2019_10_10_220000_create_promociones_menus_table.php
@@ -19,9 +19,9 @@ class CreatePromocionesMenusTable extends Migration
             $table->foreign('Promocion_Id')->references('id')->on('promociones');
             $table->unsignedBigInteger('Item_Id'); // Id de la tabla Menu
             $table->foreign('Item_Id')->references('id')->on('menus');
-            $table->float('Precio_Venta_Promo');
-            $table->float('Cantidad');
-            $table->float('Subtotal');
+            $table->decimal('Precio_Venta_Promo', 20, 2);
+            $table->decimal('Cantidad', 20, 3);
+            $table->decimal('Subtotal', 20, 2);
             $table->text('Descripcion');
             $table->string('Campo_Extra_1');
             $table->string('Campo_Extra_2');
