+++ database/migrations/2019_10_10_220236_create_promociones_articulos_table.php
@@ -19,9 +19,9 @@ class CreatePromocionesArticulosTable extends Migration
             $table->foreign('Promocion_Id')->references('id')->on('promociones');
             $table->unsignedBigInteger('Articulo_Id'); // Id de la tabla Articulos
             $table->foreign('Articulo_Id')->references('id')->on('articulos');
-            $table->float('Precio_Venta_Promo');
-            $table->float('Cantidad');
-            $table->float('Subtotal');
+            $table->decimal('Precio_Venta_Promo', 20, 2);
+            $table->decimal('Cantidad', 20, 3);
+            $table->decimal('Subtotal', 20, 2);
             $table->text('Descripcion');
             $table->string('Campo_Extra_1');
             $table->string('Campo_Extra_2');
