+++ database/migrations/2019_10_10_220419_create_servicios_table.php
@@ -16,7 +16,7 @@ class CreateServiciosTable extends Migration
         Schema::create('servicios', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->string('Nombre_Servicio');
-            $table->float('Monto_Servicio');
+            $table->decimal('Monto_Servicio', 20, 2);
             $table->text('Descripcion');
             $table->string('Campo_Extra_1');
             $table->string('Campo_Extra_2');
