+++ database/migrations/2019_10_10_220955_create_ordenes_menus_table.php
@@ -19,7 +19,7 @@ class CreateOrdenesMenusTable extends Migration
             $table->foreign('Orden_Id')->references('id')->on('ordenes');
             $table->unsignedBigInteger('Item_Id')->nullable();
             $table->foreign('Item_Id')->references('id')->on('menus');
-            $table->float('Cantidad');
+            $table->decimal('Cantidad', 20, 3);
             $table->enum('Estado', ['CANCELADA', 'EN PÉRDIDA']);
             $table->text('Descripcion');
             $table->string('Campo_Extra_1');
