+++ database/migrations/2019_10_10_221822_create_facturas_table.php
@@ -26,7 +26,7 @@ class CreateFacturasTable extends Migration
             $table->foreign('Cliente_Id')->references('id')->on('clientes');
             $table->dateTime('Fecha_Emision');
             $table->enum('Estado', ['EMITIDA', 'ANULADA']);
-            $table->float('Total');
+            $table->decimal('Total', 20, 2);
             $table->text('Descripcion')->nullable();
             $table->string('Campo_Extra_1')->nullable();
             $table->string('Campo_Extra_2')->nullable();
